requirements.txt
pyyaml>=6.0
torch>=2.0
pandas>=2.0
numpy>=1.24
